move new options class to Avalonia.Controls
